+++ CHANGELOG.md
@@ -23,6 +23,7 @@ title: Changelog
 * `<cc-badge>`: add skeleton mode
 * `<cc-input-text>`: remove the unique id generation technique and rely on Shadow DOM isolation instead.
 * `<cc-input-number>`: remove the unique id generation technique and rely on Shadow DOM isolation instead.
+* `<cc-select>`: remove the unique id generation technique and rely on Shadow DOM isolation instead.
 * New component:
   * `<cc-action-dispatcher>`